Refactor: Clean code using Layered MVC
